backend/pyproject.toml
[project]
name = "core-dispatch-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "feedgen>=1.0.0",
    "httpx>=0.28.1",
    "icalendar>=7.0.3",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "resend>=2.27.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.9",
]

[tool.pyright]
typeCheckingMode = "basic"
